+++ R/poll.R
@@ -1,17 +1,17 @@
 
 #' Poll for process I/O or termination
 #'
-#' Wait until one of the specified processes produce standard output
-#' or error, terminates, or a timeout occurs.
+#' Wait until one of the specified connections or processes produce
+#' standard output or error, terminates, or a timeout occurs.
 #'
 #' @section Explanation of the return values:
 #' * `nopipe` means that the stdout or stderr from this process was not
 #'   captured.
-#' * `ready` means that stdout or stderr from this process are ready to
-#'   read from. Note that end-of-file on these outputs also triggers
-#'   `ready`.
-#' * timeout`: the processes are not ready to read from and a timeout
-#'   happened.
+#' * `ready` means that the connection or the stdout or stderr from this
+#'   process are ready to read from. Note that end-of-file on these
+#'   outputs also triggers `ready`.
+#' * timeout`: the connections or processes are not ready to read from
+#'   and a timeout happened.
 #' * `closed`: the connection was already closed, before the polling
 #'   started.
 #' * `silent`: the connection is not ready to read from, but another
@@ -25,17 +25,18 @@
 #' then you will not be notified when it exits. If you want to wait for
 #' just a single process to end, it can be done with the `$wait()` method.
 #'
-#' @param processes A list of `process` objects to wait on. If this is a
-#'   named list, then the returned list will have the same names. This
-#'   simplifies the identification of the processes. If an empty list,
-#'   then the
+#' @param processes A list of connection objects or`process` objects to
+#'   wait on. (They can be mixed as well.) If this is a named list, then
+#'   the returned list will have the same names. This simplifies the
+#'   identification of the processes.
 #' @param ms Integer scalar, a timeout for the polling, in milliseconds.
 #'   Supply -1 for an infitite timeout, and 0 for not waiting at all.
-#' @return A list of character vectors of length two. There is one list
-#'   element for each process, in the same order as in the input list.
-#'   The character vectors' elements are named `output` and `error` and
-#'   their possible values are: `nopipe`, `ready`, `timeout`, `closed`,
-#'   `silent`. See details about these below.
+#' @return A list of character vectors of length one or two.
+#'   There is one list element for each process, in the same order as in
+#'   the input list. For connections the result is a single string scalar.
+#'   For processes the character vectors' elements are named `output` and
+#'   `error`. Possible values for each individual result are: `nopipe`,
+#'   `ready`, `timeout`, `closed`, `silent`. See details about these below.
 #'
 #' @export
 #' @examples
@@ -73,19 +74,21 @@
 #' }
 
 poll <- function(processes, ms) {
-  assert_that(is_list_of_processes(processes))
+  assert_that(is_list_of_pollables(processes))
   assert_that(is_integerish_scalar(ms))
   if (length(processes) == 0) {
     return(structure(list(), names = names(processes)))
   }
-  statuses <- lapply(processes, function(p) {
+  conn <- vapply(processes, is_connection, logical(1))
+  processes[!conn] <- lapply(processes[!conn], function(p) {
     p$.__enclos_env__$private$status
   })
 
-  res <- lapply(
-    .Call(c_processx_poll, statuses, as.integer(ms)),
-    function(x) structure(poll_codes[x], names = c("output", "error"))
-  )
-
-  structure(res, names = names(processes))
+  res <- .Call(c_processx_poll, processes, conn, as.integer(ms))
+  res <- lapply(res, function(x) poll_codes[x])
+  res[!conn] <- lapply(res[!conn], function(x) {
+    set_names(x, c("output", "error"))
+  })
+  names(res) <- names(processes)
+  res
 }